+++ internal/elasticsearch/index/index_test.go
@@ -65,6 +65,7 @@ func TestAccResourceIndexSettings(t *testing.T) {
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index.test_settings", "shard_check_on_startup", "false"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index.test_settings", "sort_field.0", "sort_key"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index.test_settings", "sort_order.0", "asc"),
+					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index.test_settings", "mapping_coerce", "true"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index.test_settings", "auto_expand_replicas", "0-5"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index.test_settings", "search_idle_after", "30s"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index.test_settings", "refresh_interval", "10s"),
@@ -249,6 +250,7 @@ resource "elasticstack_elasticsearch_index" "test_settings" {
   shard_check_on_startup = "false"
   sort_field = ["sort_key"]
   sort_order = ["asc"]
+  mapping_coerce = true
   auto_expand_replicas =  "0-5"
   search_idle_after = "30s"
   refresh_interval = "10s"